backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
sqlalchemy>=2.0.28
psycopg2-binary>=2.9.9
pydantic>=2.6.4
python-dotenv>=1.0.1
